# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)

    HTTP API to be implemented by Event Transmitters. This API can be used by
    Event Receivers to query and update the Event Stream configuration and
    status, to add and remove subjects, and to trigger verification.
  license:
    name: MIT
  title: Stream Management API for OpenID Shared Security Events
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /.well-known/sse-configuration/{issuer}:
    get:
      description: Return Transmitter Configuration information (with support for
        specifying an issuer).
      operationId: _well_known_ssf_configuration_issuer_get
      parameters:
      - description: |
          Using path components enables supporting multiple issuers per host.
          This is required in some multi-tenant hosting configurations.
          This use of .well-known is for supporting multiple issuers per host;
          unlike its use in [RFC5785](https://openid.net/specs/openid-sse-framework-1_0.html#RFC5785),
          it does not provide general information about the host.
        example: tenant-a
        explode: false
        in: path
        name: issuer
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransmitterConfiguration'
          description: |
            A set of Claims about the Transmitter's configuration,
            including all necessary endpoints and public key location information.
      summary: Transmitter Configuration Request (with path)
      tags:
      - StreamManagement
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
...
components:
...
  schemas:
...
    TransmitterConfiguration:
      description: |
        Transmitters have metadata describing their configuration.
        [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#discovery-meta)
      example:
        delivery_methods_supported:
        - https://schemas.openid.net/secevent/risc/delivery-method/push
        - https://schemas.openid.net/secevent/risc/delivery-method/poll
        configuration_endpoint: https://transmitter.most-secure.com/stream
        critical_subject_members:
        - tenant
        - user
        add_subject_endpoint: https://transmitter.most-secure.com/add-subject
        remove_subject_endpoint: https://transmitter.most-secure.com/remove-subject
        status_endpoint: https://transmitter.most-secure.com/status
        jwks_uri: https://transmitter.most-secure.com/jwks.json
        issuer: https://most-secure.com
        verification_endpoint: https://transmitter.most-secure.com/verification
      properties:
        issuer:
          description: |
            URL using the https scheme with no query or fragment component that the Transmitter asserts as its
            Issuer Identifier.
            This MUST be identical to the iss claim value in Security Event Tokens issued from this Transmitter.
          example: https://most-secure.com
          format: uri
          type: string
        jwks_uri:
          description: |
            URL of the Transmitter's [JSON Web Key Set](https://openid.net/specs/openid-sse-framework-1_0.html#RFC7517)
            document. This contains the signing key(s) the Receiver uses to validate signatures from the Transmitter.
          example: https://transmitter.most-secure.com/jwks.json
          format: uri
          type: string
        delivery_methods_supported:
          description: List of supported delivery method URIs. Recommended.
          example:
          - https://schemas.openid.net/secevent/risc/delivery-method/push
          - https://schemas.openid.net/secevent/risc/delivery-method/poll
          items:
            format: uri
            type: string
          type: array
        configuration_endpoint:
          description: The URL of the Configuration Endpoint.
          example: https://transmitter.most-secure.com/stream
          format: uri
          type: string
        status_endpoint:
          description: The URL of the Status Endpoint.
          example: https://transmitter.most-secure.com/status
          format: uri
          type: string
        add_subject_endpoint:
          description: The URL of the Add Subject Endpoint.
          example: https://transmitter.most-secure.com/add-subject
          format: uri
          type: string
        remove_subject_endpoint:
          description: The URL of the Remove Subject Endpoint.
          example: https://transmitter.most-secure.com/remove-subject
          format: uri
          type: string
        verification_endpoint:
          description: The URL of the Verification Endpoint.
          example: https://transmitter.most-secure.com/verification
          format: uri
          type: string
        critical_subject_members:
          description: |
            List of member names in a Complex Subject which, if present in a Subject Member in an event,
            MUST be interpreted by a Receiver.
          example:
          - tenant
          - user
          items:
            type: string
          type: array
      required:
      - issuer
      - jwks_uri
      title: Transmitter Configuration Response
      type: object
...
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http